refactor(rpc): isolate debug http handlers
